+++ default.nix
@@ -16,7 +16,8 @@ llvmPackages_13.stdenv.mkDerivation rec {
   version = kimeVersion;
   cargoDeps = rustPlatform.fetchCargoTarball {
     inherit src;
-    sha256 = "sha256-AguoAulVtzw6S6p+Smai+71O1ncXBob4woM1oUFCnWo=";
+    # sha256 = "0000000000000000000000000000000000000000000000000000";
+    sha256 = "sha256-XvL8VKbAqKYk63f2kz82c/QZmq/vh/nnmdnBKUAT3as=";
   };
   LIBCLANG_PATH = "${pkgs.llvmPackages_13.libclang.lib}/lib";
   dontUseCmakeConfigure = true;