+++ src/engine/backend/src/input_result.rs
@@ -5,6 +5,7 @@ bitflags::bitflags! {
         const LANGUAGE_CHANGED = 0b10;
         const HAS_PREEDIT = 0b100;
         const HAS_COMMIT = 0b1000;
+        const NOT_READY = 0b10000;
     }
 }