+++ src/engine/capi/src/lib.rs
@@ -23,9 +23,25 @@ impl RustStr {
 }
 
 #[repr(C)]
-pub struct XimPreeditFont {
-    name: RustStr,
-    size: f64,
+pub struct RustSlice {
+    ptr: *const u8,
+    len: usize,
+}
+
+impl RustSlice {
+    pub fn new(s: &[u8]) -> Self {
+        Self {
+            ptr: s.as_ptr(),
+            len: s.len(),
+        }
+    }
+}
+
+#[repr(C)]
+pub struct FontData {
+    font_data: RustSlice,
+    index: u32,
+    size: f32,
 }
 
 /// Return API version
@@ -57,6 +73,18 @@ pub extern "C" fn kime_engine_set_input_category(
 #[no_mangle]
 pub unsafe extern "C" fn kime_engine_delete(engine: &mut InputEngine) {
     drop(Box::from_raw(engine));
+}
+
+/// Check engine ready state
+#[no_mangle]
+pub unsafe extern "C" fn kime_engine_check_ready(engine: &InputEngine) -> bool {
+    engine.check_ready()
+}
+
+/// End engine ready state
+#[no_mangle]
+pub unsafe extern "C" fn kime_engine_end_ready(engine: &mut InputEngine) -> InputResult {
+    engine.end_ready()
 }
 
 /// Update layout state
@@ -146,18 +174,28 @@ pub unsafe extern "C" fn kime_config_delete(config: *mut Config) {
     drop(Box::from_raw(config));
 }
 
-/// Get xim_preedit_font config
-/// name only valid while config is live
-///
-/// ## Return
-///
-/// utf-8 string when len
+/// Get candidate_font config
+/// font_data only valid while config is live
 #[no_mangle]
-pub extern "C" fn kime_config_xim_preedit_font(config: &Config) -> XimPreeditFont {
-    let (ref font, size) = config.xim_preedit_font;
+pub extern "C" fn kime_config_candidate_font(config: &Config) -> FontData {
+    let (ref font, index) = config.candidate_font;
 
-    XimPreeditFont {
-        name: RustStr::new(font),
+    FontData {
+        font_data: RustSlice::new(font),
+        index,
+        size: 0.,
+    }
+}
+
+/// Get xim_preedit_font config
+/// font_data only valid while config is live
+#[no_mangle]
+pub extern "C" fn kime_config_xim_preedit_font(config: &Config) -> FontData {
+    let (ref font, index, size) = config.xim_preedit_font;
+
+    FontData {
+        font_data: RustSlice::new(font),
+        index,
         size,
     }
 }